==> code/1_new_data_prep.ipynb <==
RAW DATA PRE-PROCESSING

'''
Read raw biomarker data from csv, flatten by collection date
re-arrange the biomarkers in suitable date: day 1 = admission date
mark conflicting, missing data for manual correction
'''
import numpy as np
import pandas as pd

data_path = "../../data/"
raw_data = data_path + "updated_rawdata_2024_v22.xlsx"
orig_data = data_path + "All CENTER TRAJ data (to share) 6-3-19 v2.xlsx"
tony_462 = data_path + "462_samples_with_adm_date_updated_v2.csv"

validation_results = '../../results/validation/new_data02.csv'
validation_missing_data = '../../results/validation/new_missing_data1.csv'

rawdf = pd.read_excel(raw_data, sheet_name="Sheet1", header=0)
orgdf = pd.read_excel(orig_data, sheet_name="adjusted", header=0)
tdf = pd.read_csv(tony_462, header = 0)

target_col_name = tdf.columns.tolist()
target_data_type = tdf.dtypes

new_col_name = ['extra', 'Id', 'NSE', 'Date','Sample', 'S100B', 'GFAP','Tau', 'UCH-L1', 'NFL']
biomarkers = ['NSE', 'S100B', 'GFAP','Tau', 'UCH-L1', 'NFL']

df = rawdf[rawdf['subjectId'].isin(tdf['subjectId_x'])]

subjects = df['subjectId'].unique()
df.columns = new_col_name
df.loc[:, 'extra'] = None

dftest = df[df.loc[:, 'Id'] == '2AAC636']
print(dftest)
min_idx = dftest['Date'].dropna().idxmin()
print(dftest.loc[min_idx, 'NSE'])
fst_day = dftest.loc[min_idx, 'Date'] + pd.Timedelta(days=2)
if dftest['Date'].isin([fst_day]).any():
    nx_idx = dftest['Date'].index[dftest['Date'] == fst_day]
    print(dftest.loc[nx_idx, 'GFAP'])
else:
    print('Not found')



value_map = {
    'Collection_date_'  : 'Date',
    'sampleid_'         : 'Sample',
    'NSE_'              : 'NSE',
    'S100B_'            : 'S100B',
    'GFAP_'             : 'GFAP',
    'Tau_'              : 'Tau',
    'UCH-L1_'           : 'UCH-L1',
    'NFL_'              : 'NFL'
}
dfx = pd.DataFrame(columns=target_data_type.index)
for i, id in enumerate(subjects):
    dfi = df[df.loc[:, 'Id'] == id]
    # print('dfi = \n', dfi)
    nday = dfi['Date'].dropna().min()
    # row = {}
    dfx.loc[i, 'subjectId_x'] = id
    counter = 0
    for j in range(min(5, len(dfi))):
        if counter > 4:
            break
        
        if dfi['Date'].isin([nday]).any():
            # idx_day = dfi['Date'].isin([nday]).index[0]
            idx_days = dfi[dfi['Date'] == nday].index
            if (len(idx_days) > 1):
                dfx.loc[i, 'Manual_Review'] = 'B'
            for ix, idx_day in enumerate(idx_days):
                col_suffix = str(counter + 1)
                for k, v in value_map.items():
                    dfx.loc[i, k + col_suffix] = dfi.loc[idx_day, v]
                counter += 1
                if counter > 4:
                    break
        else:
            counter += 1
        nday = nday + pd.Timedelta(days=1)                        
            
        
            
    if (counter < 5 and dfi['Date'].isnull().any()):
        dfx.loc[i, 'Manual_Review'] = 'A'
print(dfx.shape)

    

print(dfx)

# Export to file for manual correction
print(dfx)

dfx.to_csv(validation_results, header=True, index=False)

# Data that need to update manually selected 18 rows
missing_data_id = dfx[dfx['Manual_Review'] == 'A']['subjectId_x']
dfy = df[df['Id'].isin(missing_data_id)]
dfy = dfy[dfy['Date'].isnull()]
dfy.to_csv(validation_missing_data, header=True, index=False)
print(dfy)
